+++ agent/tandem/agent/io/document.py
@@ -1,7 +1,7 @@
 import json
 import os
 from subprocess import Popen, PIPE
-from tandem.configuration import CRDT_PATH
+from tandem.agent.configuration import CRDT_PATH
 
 CRDT_PROCESS = ["node", os.path.join(CRDT_PATH, "build", "bundle.js")]
