+++ agent/tandem/agent/protocol/messages/editor.py
