+++ rendezvous/tandem/rendezvous/protocol/handlers/agent.py
@@ -1,7 +1,7 @@
 from tandem.rendezvous.models.connection import Connection
 from tandem.rendezvous.stores.connection import ConnectionStore
 from tandem.rendezvous.stores.session import SessionStore
-from tandem.shared.protocol.handler.base import ProtocolHandlerBase
+from tandem.shared.protocol.handlers.base import ProtocolHandlerBase
 from tandem.shared.protocol.messages.rendezvous import (
     RendezvousProtocolMessageType,
     RendezvousProtocolUtils,
